chor: refactor `tests/` directories and files
