+++ tests/figure.test.ts
@@ -1,5 +1,5 @@
 import { stripIndent } from 'common-tags';
-import { buildProcessorTestingCode } from '../utils';
+import { buildProcessorTestingCode } from './utils';
 
 it(
   'simple figure output',
