+++ tests/line-breaks.test.ts
@@ -1,5 +1,5 @@
 import { stripIndent } from 'common-tags';
-import { buildProcessorTestingCode } from '../utils';
+import { buildProcessorTestingCode } from './utils';
 
 // remark-parse GFM seems to support only EOL two-spaces, so backslash is excluded.
 it(
